@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -47,25 +47,12 @@ dependencies:
   connectivity: 3.0.6
   device_info: 2.0.3
   url_launcher: 6.0.17
-  google_fonts: 2.1.0
   android_intent: 2.0.2
   shared_preferences: 2.0.10
-  auto_size_text: 3.0.0
   flutter_localizations:
     sdk: flutter
-  expandable: ^5.0.1
-  gradient_ui_widgets: ^1.0.2-nullsafety
-  modal_bottom_sheet: ^2.0.0
-  # tencent_im_sdk_plugin: 3.5.6
-  flutter_syntax_view: ^4.0.0
   loading_indicator: ^2.0.1
   crypto: ^3.0.1
-  flutter_easyloading: ^3.0.0
-  archive: ^3.1.6
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  pretty_json: ^2.0.0
-  adaptive_action_sheet: ^2.0.0
   image_picker: ^0.8.4
   video_thumbnail: ^0.4.3
   record: ^2.1.1
@@ -80,51 +67,13 @@ dependencies:
   #  cached_network_image: 1.1.1 # 1.9版本flutter
   #  cached_network_image: 2.0.0 #1.22.5版本Flutter用这个
   cached_network_image: ^3.2.0
-  #toast提示框
-  toast: 0.1.5
-  #文本扩展显示
-  #  extended_text: 1.0.1  #1.17版本Flutter用这个
-  #  extended_text: 4.0.0 #1.20版本Flutter用这个
-  #  extended_text: 4.0.1 #1.22.5版本Flutter用这个
-  #  extended_text: ^6.0.5 #Flutter2用这个,,其他可注释
-  extended_text: ^9.0.0
-  #文字表情混入
-  #  extended_text_field: 1.0.1  #1.17版本Flutter用这个
-  #  extended_text_field: 4.0.0  #1.20版本Flutter用这个
-  #  extended_text_field: 5.0.1 #1.22.5版本Flutter用这个
-  #  extended_text_field: ^7.0.1 #Flutter2用这个,,其他可注释
-  extended_text_field: ^10.0.0
-  #  extended_text_library: 4.0.2 #1.22.5版本Flutter用这个,,其他可注释
-  #  extended_text_library: ^6.0.0 #Flutter2用这个,,其他可注释
-  extended_text_library: ^9.0.0
-  #打开系统浏览器
-  #网络连接
   #图片显示
   photo_view: ^0.13.0
-  #  dio: 2.2.2  #低于Flutter-1.17使用此版本
-  #  dio: 3.0.9  #1.22.5版本Flutter用这个
-  #打开文件
-  #权限处理
-  #音频处理
-  #  audioplayers: 0.13.3 #1.22.5版本Flutter用这个
-  #调用相机
-
-  #即时通讯【腾讯云】
-  dim:
-    path: plugin/wechat_flutter_plugin
-  #持久化存储
-  #打开文件
-  open_file: ^3.2.1
-  #获取包信息
-  #音频处理
   flutter_sound: ^9.1.3
   #权限处理
   #音频处理
   #  audioplayers: 0.13.3 #1.22.5版本Flutter用这个
   audioplayers: ^0.20.1
-  #调用相机
-  camera: # 0.5.7+1
-    path: plugin/camera
   #视频播放
   #  video_player: 0.10.11+2 #1.22.5版本Flutter用这个
   video_player: ^2.2.19
