@@ -1,18 +1,15 @@
-import 'dart:collection';
-
 import 'package:doctor_project/common/style/gsy_style.dart';
 import 'package:doctor_project/pages/home/clinical_reception_person_set.dart';
 import 'package:doctor_project/pages/home/submit_success.dart';
 import 'package:doctor_project/pages/home/topic_price_set.dart';
 import 'package:doctor_project/utils/colors_utils.dart';
 import 'package:doctor_project/widget/custom_app_bar.dart';
-import 'package:doctor_project/widget/gsy_flex_button.dart';
 import 'package:doctor_project/widget/safe_area_button.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class HealthConsultService extends StatefulWidget {
-  const HealthConsultService({Key key}) : super(key: key);
+  const HealthConsultService({Key? key}) : super(key: key);
 
   @override
   _HealthConsultServiceState createState() => _HealthConsultServiceState();
