@@ -1,5 +1,4 @@
 import 'package:doctor_project/common/style/gsy_style.dart';
-import 'package:doctor_project/tools/wechat_flutter.dart';
 import 'package:doctor_project/utils/colors_utils.dart';
 import 'package:doctor_project/widget/custom_app_bar.dart';
 import 'package:doctor_project/widget/safe_area_button.dart';
@@ -7,7 +6,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class CreateTemplate extends StatefulWidget {
-  const CreateTemplate({Key key}) : super(key: key);
+  const CreateTemplate({Key? key}) : super(key: key);
 
   @override
   _CreateTemplateState createState() => _CreateTemplateState();
@@ -70,7 +69,7 @@ class _CreateTemplateState extends State<CreateTemplate> {
                           '请选择科室',
                           style: GSYConstant.textStyle(color: '#999999'),
                         ),
-                        Space(
+                        SizedBox(
                           width: 8.0,
                         ),
                         Image.asset('assets/images/my/forward.png')
