@@ -7,7 +7,7 @@ class PlatformUtils {
   /// 返回值或运行基于平台的函数。
   /// 如果传递了上下文，它将通过 Theme.of(context).platform 获取平台。
   /// 否则，它将使用 defaultTargetPlatform。
-  static get({@required BuildContext context}) {
+  static get({required BuildContext context}) {
     return context != null ? Theme.of(context).platform : defaultTargetPlatform;
   }
   static bool isIPhoneX(BuildContext context) {
@@ -16,7 +16,7 @@ class PlatformUtils {
     }
     return false;
   }
-  static T select<T>({@required BuildContext context, dynamic android, dynamic ios,
+  static T select<T>({required BuildContext context, dynamic android, dynamic ios,
     dynamic fuchsia, dynamic web, dynamic macOS,dynamic windows,
     dynamic linux,dynamic defaultWhenNull}) {
     var func;
