@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
 class CustomExpansionPanelList extends ExpansionPanelList{
-  const CustomExpansionPanelList({Key key}) : super(key: key);
+  const CustomExpansionPanelList({Key? key}) : super(key: key);
 
 }
